Refactor Employee class to remove 'id' field
The 'id' field has been removed from the Employee class and associated methods. The 'employee_id' field is now of type int and will be used as the unique identifier. This change simplifies the structure and reduces redundancy as both 'id' and 'employee_id' fields previously served as identifiers. Update on SQL queries have been made accordingly to reflect this change.